@@ -1,10 +1,10 @@
 package com.criticalblue.android.astropiks;
 
 import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.LayoutRes;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * Provides container for single fragment activities.
@@ -34,5 +34,3 @@ public abstract class SingleFragmentActivity extends AppCompatActivity {
         }
     }
 }
-
-// end of file
