@@ -14,36 +14,22 @@
  * limitations under the License.
  */
 
-const path = require('path');
 const url = require('url');
 const request = require('request');
-const chalk = require('chalk');
+const log = require('./../logging')
 
 // load api configuration and secrets
-
 const config = require(`${__dirname}/../config.js`);
-
-if (config.nasa_host == null) {
-  throw new Error(`nasa_host not found; please set in ${__dirname}/../config.js`);
-}
-const api_host = config.nasa_host;
-
-if (config.nasa_protocol == null) {
-  throw new Error(`nasa_protocol not found; please set in ${__dirname}/../config.js`);
-}
-const api_protocol = config.nasa_protocol;
-
-const secrets = require(`${__dirname}/../secrets.js`);
-
-if (secrets.nasa_api_key == null) {
-  throw new Error(`nasa_api_key not found; please set in ${__dirname}/../secrets.js`);
-}
-const api_key = secrets.nasa_api_key;
+const api_host = config.NASA_HOST;
+const api_protocol = config.NASA_PROTOCOL;
+const nasa_api_key = config.NASA_API_KEY;
+const proxy_protocol = 'https'
 
 const apodHostname = 'apod.nasa.gov';
 const apodRoute = '/' + apodHostname + '/apod/image/*'
-const apodDirect = api_protocol + '//' + apodHostname + '/';
+const apodDirect = api_protocol + '://' + apodHostname + '/';
 const apodDirectRe = new RegExp(apodDirect.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&'), 'g');
+
 
 /**
  * Describes NASA API route handlers.
@@ -53,37 +39,37 @@ const apodDirectRe = new RegExp(apodDirect.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\
 function routes(app) {
 
   // proxy a picture of the day request
-
   app.use(`/${api_host}`, (req, res, next) => {
-    console.log('Processing NASA API request');
+    log.info('Start Processing the NASA API Request...');
 
     // build redirected request
-
-    var urlInfo = url.parse(req.url, true);  // build proxied url
+    let urlInfo = url.parse(req.url, true);
     urlInfo.protocol = api_protocol;
     urlInfo.host  = api_host;
-    // urlInfo.pathname is req.url
+    urlInfo.query.api_key = nasa_api_key;
     delete urlInfo.search;
-    urlInfo.query.api_key = api_key;         // add nasa api key
-    var nasaUrl = url.format(urlInfo);
 
-    var proxyHome = req.protocol + '://' + req.headers.host;
-    var apodProxy = proxyHome + '/' + apodHostname + '/';
+    let nasaUrl = url.format(urlInfo);
+    log.warning("URL: " + nasaUrl)
 
-    var nasaHdrs = req.headers;              // reuse most headers
+    let proxyHome = proxy_protocol + '://' + req.headers.host;
+    let apodProxy = proxyHome + '/' + apodHostname + '/';
+
+    // reuse most headers
+    let nasaHdrs = req.headers;
     delete nasaHdrs['host'];
     delete nasaHdrs['accept-encoding'];
 
     // start proxy request
-
     request({ url: nasaUrl, headers: nasaHdrs }, (err, proxyRes, proxyBody) => {
       if (err) {
-        console.log(chalk.red(`Internal Server Error: in NASA proxy: ${err}`));
+        log.fatalError(`Internal Server Error: in NASA proxy: ${err}`);
         res.status(500).send('Internal Server Error');
       } else {
 
-        // patch response to redirect any apod image requests through proxy
+        log.info("Replacing NASA apod image url in the response body...")
 
+        // patch response to redirect any apod image requests through proxy
         proxyBody = proxyBody.replace(apodDirectRe, apodProxy);
         proxyRes.headers["content-length"] = proxyBody.length.toString();
 
@@ -95,21 +81,16 @@ function routes(app) {
   });
 
   // proxy a picture of the day image download
-
   app.get(apodRoute, (req, res, next) => {
-    console.log('Processing NASA apod image request', api_protocol + '/' + req.url);
 
-    // console.log('image headers:', JSON.stringify(req.headers, null, '  '));
+    let proxyUrl = api_protocol + ':/' + req.url;
+    log.warning('Processing NASA apod image request for: ' + proxyUrl);
 
-    // pipe the image request through the proxy
-
-    let proxyUrl = api_protocol + '/' + req.url;
     let proxyReq = request(proxyUrl);
 
+    // pipe the image request through the proxy
     req.pipe(proxyReq).pipe(res);
   });
 }
 
 module.exports = { routes: routes };
-
-// end of file
